Twekaed: VelvetNoiseReverb: Improved low and high pass filters frequency settings

+++ Shared/Utilities/NoiseGenerator.h
@@ -194,7 +194,7 @@ public:
 		}
 		else
 		{
-			return m_pieceWiceDistribution(m_mersenneTwisterGenerator);
+			return (float)(m_pieceWiceDistribution(m_mersenneTwisterGenerator));
 		}
 	}
